+++ runner/src/run.rs
@@ -1,42 +1,21 @@
 // use shared::uv_handler::find_or_download_uv;
+use shared::uv_handler::find_or_download_uv;
+use shared::{debug_println, debuging};
 use std::path::Path;
 use std::path::PathBuf;
 use std::process::Command;
 use std::{self, io};
-use shared::uv_handler_v2::find_or_download_uv;
-use shared::{debug_println, debuging};
 
 use shared::config::{ProjectConfig, load_project_config};
 
-fn find_manifest_file(project_dir: &Path) -> io::Result<PathBuf> {
-    let manifest_files = [
-        "pyproject.toml",
-        "requirements.txt",
-        "pylock.toml",
-        "setup.py",
-        "setup.cfg",
-    ];
-
-    for file in &manifest_files {
-        let path = project_dir.join(file);
-        if path.exists() {
-            return Ok(path);
-        }
-    }
-
-    Err(io::Error::new(
-        io::ErrorKind::NotFound,
-        "No manifest file found in the source directory. \nManifest files can be pyproject.toml, requirements.txt, pylock.toml, setup.py or setup.cfg",
-    ))
-}
-
 fn apply_env_from_config(config: &ProjectConfig) {
     if let Some(env_config) = &config.env
-        && let Some(vars) = &env_config.variables {
-            for (k, v) in vars {
-                unsafe { std::env::set_var(k, v) }; // Set env variables - not thread safe
-            }
+        && let Some(vars) = &env_config.variables
+    {
+        for (k, v) in vars {
+            unsafe { std::env::set_var(k, v) }; // Set env variables - not thread safe
         }
+    }
 }
 
 fn prepare_hooks(config: &ProjectConfig) -> (String, String) {
@@ -56,12 +35,7 @@ fn prepare_hooks(config: &ProjectConfig) -> (String, String) {
     (pre_hook, post_hook)
 }
 
-fn run_uv(
-    uv_path: &Path,
-    project_dir: &Path,
-    with: &[&str],
-    args: &[&str],
-) -> io::Result<()> {
+fn run_uv(uv_path: &Path, project_dir: &Path, with: &[&str], args: &[&str]) -> io::Result<()> {
     let mut cmd = Command::new(uv_path);
     cmd.arg("run").arg("-q");
 
@@ -69,8 +43,7 @@ fn run_uv(
         cmd.arg("--with").arg(w);
     }
 
-    cmd.args(args)
-       .current_dir(project_dir);
+    cmd.args(args).current_dir(project_dir);
 
     let status = cmd.status()?;
 
@@ -104,10 +77,10 @@ fn find_single_wheel(project_dir: &Path) -> io::Result<Option<PathBuf>> {
     Ok(wheel)
 }
 
-
 pub fn run_extracted_project(project_dir: &Path, runtime_args: Vec<String>) -> io::Result<()> {
     // Load project configuration and determine entrypoint
     let config = load_project_config(&project_dir.to_path_buf());
+    debug_println!("[main.run_extracted_project] - Loaded project configuration");
 
     // Enable debug mode if specified in config
     if config.options.debug {
@@ -117,10 +90,11 @@ pub fn run_extracted_project(project_dir: &Path, runtime_args: Vec<String>) -> i
 
     // Ensure UV is available
     debug_println!("[main.run_extracted_project] - Ensuring UV is available");
-    let uv_path = find_or_download_uv(None).ok_or(io::Error::new(
-        io::ErrorKind::NotFound,
-        "Could not find or download uv binary",
-    ))?;
+    let uv_path =
+        find_or_download_uv(None, config.options.uv_version.as_str()).ok_or(io::Error::new(
+            io::ErrorKind::NotFound,
+            "Could not find or download uv binary",
+        ))?;
 
     // Apply environment variables from config (unsafe but we are single-threaded so it should be fine)
     apply_env_from_config(&config);
@@ -129,7 +103,6 @@ pub fn run_extracted_project(project_dir: &Path, runtime_args: Vec<String>) -> i
     );
 
     // Determine entrypoint
-    debug_println!("[main.run_extracted_project] - Loaded project configuration");
     let entrypoint = &config.package.entrypoint;
     let entry_point_path = project_dir.join(entrypoint);
     if entrypoint.ends_with(".py") {
@@ -164,11 +137,12 @@ pub fn run_extracted_project(project_dir: &Path, runtime_args: Vec<String>) -> i
     );
 
     // Run pre-hook if specified
-    debug_println!("[main.run_extracted_project] - Running pre-hook if specified");
     if !pre_hook.is_empty() {
+        debug_println!("[main.run_extracted_project] - Running pre-hook");
         run_uv(&uv_path, project_dir, &[], &[pre_hook.as_str()])?;
     }
 
+    debug_println!("[main.run_extracted_project] - Running main project");
     match run_mode {
         "source" => {
             debug_println!("[main.run_extracted_project] - Running in source mode");
@@ -178,33 +152,34 @@ pub fn run_extracted_project(project_dir: &Path, runtime_args: Vec<String>) -> i
 
             let args_refs: Vec<&str> = args_vec.iter().map(|s| s.as_str()).collect();
             run_uv(&uv_path, project_dir, &[], &args_refs)?;
-        },
+        }
         "wheel" => {
             debug_println!("[main.run_extracted_project] - Running in wheel mode");
             let wheel_file = wheel.ok_or(io::Error::new(
                 io::ErrorKind::NotFound,
                 "No .whl file found in the project directory",
             ))?;
-            run_uv(&uv_path, project_dir, &[wheel_file.to_str().unwrap()], &[config.package.entrypoint.as_str()])?;
-        },
+            run_uv(
+                &uv_path,
+                project_dir,
+                &[wheel_file.to_str().unwrap()],
+                &[config.package.entrypoint.as_str()],
+            )?;
+        }
         _ => unreachable!(),
     }
 
-
     // Run post-hook if specified
-    debug_println!("[main.run_extracted_project] - Running post-hook if specified");
     if !post_hook.is_empty() {
+        debug_println!("[main.run_extracted_project] - Running post-hook");
         run_uv(&uv_path, project_dir, &[], &[post_hook.as_str()])?;
     }
 
     // Clean up if delete_after_run is set or extract_to_temp is set
-    debug_println!(
-        "[main.run_extracted_project] - Cleaning up extracted project if configured to do so"
-    );
-    if (config.options.delete_after_run || config.options.extract_to_temp)
-        && project_dir.exists() {
-            std::fs::remove_dir_all(project_dir)?;
-        }
+    if (config.options.delete_after_run || config.options.extract_to_temp) && project_dir.exists() {
+        debug_println!("[main.run_extracted_project] - Cleaning up extracted project");
+        std::fs::remove_dir_all(project_dir)?;
+    }
 
     Ok(())
 }