+++ shared/src/spinner.rs
@@ -1,7 +1,6 @@
 use spinners::{Spinner, Spinners};
 
 pub fn create_spinner_with_message(msg: &str) -> Spinner {
-    
     Spinner::new(Spinners::Dots9, msg.into())
 }